+++ module-infra/security/src/main/kotlin/demo/kotlinpractice/SecurityConfig.kt
@@ -1,17 +1,4 @@
-package demo.kotlinpractice.global.config
-
-import demo.kotlinpractice.global.jwt.JwtAuthorizationFilter
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.authentication.AuthenticationManager
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.crypto.factory.PasswordEncoderFactories
-import org.springframework.security.crypto.password.PasswordEncoder
-import org.springframework.security.web.DefaultSecurityFilterChain
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+package demo.kotlinpractice
 
 @Configuration
 @EnableWebSecurity
