+++ module-infra/security/src/main/kotlin/demo/kotlinpractice/principal/AuthDetails.kt
@@ -1,8 +1,5 @@
-package demo.kotlinpractice.domain.auth
+package demo.kotlinpractice.principal
 
-import org.springframework.security.core.GrantedAuthority
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.userdetails.UserDetails
 import java.util.Collections
 
 data class AuthDetails(
